:truck: move find competence evaluations by assessment usecase
